+++ src/main/java/com/example/bankservice/account/dto/DepositRequest.java
@@ -4,8 +4,6 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigDecimal;
-
 /**
  * 입금 요청 DTO
  *
@@ -23,8 +21,8 @@ public class DepositRequest {
     private String accNumber;
 
     /**
-     * amount: 입금액
+     * amount: 입금액 (정수)
      * - 양수만 가능
      */
-    private BigDecimal amount;
+    private Long amount;
 }
