+++ src/main/java/com/example/bankservice/account/dto/WithdrawRequest.java
@@ -4,8 +4,6 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigDecimal;
-
 /**
  * 출금 요청 DTO
  *
@@ -30,9 +28,9 @@ public class WithdrawRequest {
     private String accPassword;
 
     /**
-     * amount: 출금액
+     * amount: 출금액 (정수)
      * - 양수만 가능
      * - 잔액보다 클 수 없음
      */
-    private BigDecimal amount;
+    private Long amount;
 }
